requirements.txt
numpy>=1.21
opencv-python>=4.5
albumentations
torch
torchvision
streamlit
pillow
